+++ mobile/cmp/navigator/NavigatorModel.js
@@ -26,6 +26,9 @@ export class NavigatorModel {
     /** @member {Object[]} */
     pages = [];
 
+    /** @member {boolean} */
+    track;
+
     /** @member {RenderMode} */
     renderMode;
 
@@ -49,6 +52,7 @@ export class NavigatorModel {
     /**
      * @param {Object[]} pages - configs for PageModels, representing all supported
      *      pages within this Navigator/App.
+     * @param {boolean} [track] - True to enable activity tracking of page views (default false).
      * @param {RenderMode} [renderMode] - strategy for rendering pages. Can be set per-page
      *      via `PageModel.renderMode`. See enum for description of supported modes.
      * @param {RefreshMode} [refreshMode] - strategy for refreshing pages. Can be set per-page
@@ -56,6 +60,7 @@ export class NavigatorModel {
      */
     constructor({
         pages,
+        track = false,
         renderMode = RenderMode.LAZY,
         refreshMode = RefreshMode.ON_SHOW_LAZY
     }) {
@@ -65,6 +70,7 @@ export class NavigatorModel {
         ensureUniqueBy(pages, 'id', 'Multiple NavigatorModel PageModels have the same id.');
 
         this.pages = pages;
+        this.track = track;
         this.renderMode = renderMode;
         this.refreshMode = refreshMode;
 
@@ -77,6 +83,18 @@ export class NavigatorModel {
             track: () => this.stack,
             run: this.onStackChangeAsync
         });
+
+        if (track) {
+            this.addReaction({
+                track: () => this.activePageId,
+                run: (activePageId) => {
+                    XH.track({
+                        category: 'Navigation',
+                        message: `Viewed ${activePageId} page`
+                    });
+                }
+            });
+        }
     }
 
     /**